Add unit test for Groups List dropdown controller
Add a test for the controller part of the <groups-list> directive.

We have a couple of different patterns for testing directive
controllers in the code, the approach taken here follows the privacy
dropdown's directive in simply exporting the controller's constructor
as a 'private' export and referencing it directly in the test.

 * Rename the exports in the 'directive/privacy' module for consistency