Remove remnants of annotator store auth
The consumer field is removed entirely because with authorization
disabled in annotator-store the field serves no purpose at present.

Claims used by the Annotator Auth plugin on the front end are removed
from the JWT since this client has been deprecated for some time.

The User and Consumer model stubs from annotator-store are removed
and the authenticated userid from Pyramid is passed to the logic
functions of h.api now.