Turn off HTTP syntax highlighting in API docs
The HTTP syntax highlighting doesn't work for requests/replies with JSON
bodies if the example JSON body isn't valid JSON. We don't want to give
complete example JSON bodies all the time, we want to use ... to
indicate there is more data here, but that's not valid JSON. The result
is that some requests/responses are highlighted and others aren't. So
just disable the syntax highlighting for all of them, at least it's
consistent.