Change $anchored -> $orphan
Instead of setting $anchored=true on anchored annotations, set
$orphan=true on orphaned (i.e. unanchored) ones as we used to do.

Orphaned annotations are the exception: we don't show them, or if
show_unanchored_annotations is one then we show them with a warning.
It's simpler and less buggy to add $orphan=true onto the exceptional
annotations than to try to add $anchored=true onto all of the
unexceptional ones.

Fixes #2525.
Fixes #2541.