Limit group names to 25 characters
Enforce this limit in Python (colander schema) only, in the database
leave the limit at its current value of 100 characters. This is to
avoid having to do a database migration. In the future we plan to remove
all character limits from database text columns (including this group
name column) and enforce them all in Python only, to avoid having to do
painful database migrations.