Fix broken test in h.api.groups.auth
This test was attempting to test that when the annotation group is any
of

- unset
- None
- the empty string
- '__world__'

then `set_permissions` doesn't modify the permissions field. But it
couldn't detect this because a) it was making a shallow copy so changes
to the permissions field wouldn't be detected, and b) the annotation was
private so was unchanged for a different reason.