Fix annotation privacy note moving depending on which item was selected
Due to the existence of a 'p + p' selector in app.css, the <p>
element containing the 'This annotation is visible only to you'
label was displayed at a different vertical position to
the alternative labels.

Use ng-if rather than ng-show to remove the non-visible
labels so that the rendered <p> is always the first <p> child
of its parent and therefore always rendered at the same position.