Implement the sticky privacy behavior in the new annotation publish button
Re-implement the behavior where new annotations
use the last-used privacy setting by default.

This was previously implemented in privacy.js
and the logic now lives in the replacement component,
publish-annotation-btn.js. However, the defaults
really shouldn't be set in the view but earlier when
the new annotation is created.

Card 89