Check permissions correctly when sending reply notifications
When we receive an annotation in the notification mailer, we need to
perform a set of checks:

- to determine if the annotation is a reply
- to determine who the author of the thread parent is
- to determine if we should send a reply notification to that author

In particular, we only want to send a reply notification email if the
author is going to be able to read the reply. Previously, this was done
by checking if the principals allowed by the 'read' permission included
the system principal "system.Everyone".

Unfortunately, 998347a broke this by removing normal Annotation
instances' status as a pyramid "context object", meaning that no
annotations were passing the test and no reply notification emails were
being sent.

This commit updates that permissions check. Now we:

1. Compute the set of principals of the parent annotation's author,
   using `h.auth.effective_principals`.
2. Compute the set of principals that are allowed to 'read' the reply
   annotation.
3. Abort unless these two sets have a non-null intersection.

Fixes #2555.