Make auth.effective_principals do what it says
Previously `h.auth.effective_principals` was a "groupfinder", returning
the list of groups of which a given userid is a member.

This commit renames that function to `groupfinder`, and updates the two
configuration files to use it in that capacity. It also adds a new
helper function to `h.auth`, which actually does return the list of
effective principals for a given userid.