Factor out translation of annotator->pyramid principals
Annotator uses virtual principals such as 'group:__world__', while
pyramid uses its own `system.Everyone`. We need to be able to reliably
translate a set of annotator permissions into something meaningful
inside the pyramid application, and this change factors that out of the
Annotation resource so that it can be used independently.