Warn about slow tests and add extended timeout for flakey test
Enable reporting of all tests that take longer than 500ms
to execute. In combination with high or disabled timeouts
for individual tests, this can be used to determine whether
tests that intermittently time out within the 2000ms threshold
are slow or hanging or failing to trigger the async callback.

Add a high timeout for the Bridge.createChannel() test which
is intermittently failing, eg. in https://travis-ci.org/hypothesis/h/builds/82894887

A few things to note about that run:
 - The overall completion time for the test suite was higher
   than average by more than 2000ms, so other tests may have
   taken a lot longer to run than normal
 - The test which timed out is pretty simple and synchronous,
   but there are a number of other async tests in the same suite so
   perhaps one of those tests is interfering with it

See #2563