Make activation created_by/valid_until fields nullable
We don't use these fields at all -- they are a hangover from when we
used horus for our accounts system -- and we'd like to remove them.
Unfortunately removing them in one step isn't possible with our current
stage/prod setup, because when we run the migration (in order to make
stage work) prod will break, because the fields are non-nullable and
thus will always be sent by sqlalchemy.

We can work around this by removing them in two stages:

1. Make the fields nullable. This code can be deployed to stage and the
   migration run without breaking prod. When we deploy the code to prod,
   it will then stop sending values to the database with every INSERT.

2. Remove the fields. Again, we can do this in stage and run the
   migration (to remove the columns in the database) without breaking
   prod.

This commit represents step 1.