Focus on a group
When a user clicks on a group name in the "scope selector" dropdown
menu, we want to filter the visible annotations by group. This commit
implements a rather inefficient but functional approach to this, simply
reloading the angular view.

In addition, we make sure that the initial annotation load includes a
filter on the currently-focused group, and we discard non-focused-group
annotations coming in over the websocket.