Use shims for Angular 1.2x compatibility with UI Bootstrap instead of patching upstream
Instead of patching the upstream UI Bootstrap build, add a wrapper around
it which stubs the $templateRequest service and wraps the Angular 1.2x implementation
of $animate.(add|remove)Class with one which returns a Promise as
UI Bootstrap expects.

When the app is upgraded to Angular 1.3x+, the 'ui-bootstrap-custom' module can just
be removed.

Card 89