Fix unsaved changes to annotations being lost when changing permissions
When changing permissions via the dropdown menu,
the changed permissions were applied to the _domain model_
representing the saved annotation instead of the _view model_
representing the current annotation being edited.

The change to the domain model triggered a reset of the view
model via AnnotationController.render(), losing any unsaved edits.

This commit makes AnnotationController.setPrivacy() apply permissions
changes to the view model in the same way as other edits in the
annotation view. These then only take effect when the 'Post to'
button is clicked.

Fixes #2567