Make sqlalchemy ignore old columns on INSERT/SELECT/UPDATE
These columns are going to be removed shortly. In 25ae155 we made them
nullable, in the hope that this would stop sqlalchemy from attempting to
set values in these columns. Unfortunately, that doesn't do it:
sqlalchemy continues to send explicit NULL values to these columns.

In order to get sqlalchemy to completely ignore these unused columns on
INSERT, on UPDATE and on SELECT, we have to:

- mark the `server_default` and `server_onupdate` values to be
  `FetchedValue()` -- this is usually used when a field is updated
  server-side by a trigger, and suppresses inclusion of the field in
  INSERTs and UPDATEs respectively:

  http://docs.sqlalchemy.org/en/rel_1_0/core/defaults.html#sqlalchemy.schema.FetchedValue

- mark the entire column as deferred, to prevent eager loading of these
  column values on SELECT:

  http://docs.sqlalchemy.org/en/rel_1_0/orm/loading_columns.html#deferred-column-loading