Always focus public group if groups feature off
Always focus the Public / __world__ group if the groups feature is
toggled off.

When the groups feature was off a group other than Public might have
been retrieved from the browser's localStorage (this could only happen
if the groups feature had previously been on for the user) and the wrong
annotations would be shown in the sidebar.