Show filenames of local documents with titles
When the annotated document is a local (file://) document and it has a
title, then show the filename in place of the domain on the annotation
card. For example:

  seanh on "Document Title" (filename.pdf)

In cases where the document has no title then the filename would already
have been used in place of "Document Title" so don't repeat it in
(...)'s.

For remote documents the domain name is shown in (...)'s instead and the
document title is hyperlinked so the people can see the full URL by
hovering over the link.