Insert 'group': '__world__" if none
When creating or updating an annotation if the annotation has no 'group'
field then insert 'group': '__world__'.

This fixes a bug where people calling our API directly can create
annotations with no group field, then for example when the sidebar
searches for annotations with 'group': '__world__' it won't find these
annotations, but the Chrome extension badge (which doesn't put any group
filter in its searches) will find them.

This could be fixed client-side but it's better to make the annotations
in our search index consistent, than to try and always remember that
public annotations may either have group: '__world__' or no group at
all.

This means that it's no longer necessary for us to insert
group: '__world__' on the way out in render().

Annotations pre-dating this commit with no group can be migrated with:

    hypothesis annotool conf/production.ini prepare

Fixes #2589.