Inline the H icon font
The previous fix to use an absolute URL to the icon
font did not work in the Chrome extension because
the base path for assets was different.

The proper solution here will involve refactoring
the CSS build pipeline and possibly automatically
inlining font URLs in the process. I did not just
add a preprocessor to inline all URLs because
that would also have affected fonts in katex.min.css
and added significantly to the size of the CSS bundle.

This commit reverts back to the previous solution of
inlining just the icon font.

See #2571