Remove the 'queue' feature flag
This is also deployed and toggled on everywhere.

The fact that this data is relayed through a pyramid event
(`h.api.events.AnnotationEvent`) already provides a suitable decoupling
between the our web application's use of NSQ, and the `h.api` package's
interface. By moving the code that actually publishes to NSQ into
`h.subscribers` we remove one point of dependency between `h.api` and
NSQ (the `h.api.nipsa` module still needs to be addressed).