Refactor the Chrome extension blocklist
Refactor the Chrome extension to retrieve whether the current tab's URI
is blocked or not from the server, instead of querying its own baked-in
blocklist.

Previously we had two blocklists: a list of sites that h won't launch
on, which came from the Pyramid config file and was baked into the
Chrome extension at build time. And a list of sites that the Chrome
extension won't show the number of annotations in its badge on, this was
stored in the database server-side.

This refactors the launch blocklist to use the same in-database,
server-side blocklist as used for the badge numbers.

Also remove blocklist support from the bookmarklet - the bookmarklet
just launches on every site now.

Design:

* The /badge API endpoint now returns the badge number (`total`) and
  whether or not the URI is blocked

* blocklist.js is responsible for getting the URI info from the server
  and validating it.

* hypothesis-chrome-extension.js is responsible for calling blocklist.js on
  each page load, and passing `total` into browser-action.js (updateBadge()),
  and calling or not calling injectIntoTab() (sidebar-injector.js) depending
  on whether the URI is blocked.

* All other modules know nothing about the blocklist and how it's retrieved
  from the server or comes in the form of a Promise.

  browser-action.js no longer needs to make the API request to get the
  `total` to display as it gets it passed in, sidebar-injector.js no
  longer needs to check the blocklist as it just doesn't get called on
  blocked URIs.

Changes:

blocklist.js:
New file, adds an h.getBlocklist() function that makes a request to the
server's /blocklist URI, validates the response, and returns a Promise that
resolves to the validated response.

browser-action.js:
updateBadge() now takes a `total` parameter, the number of annotations that
should be shown on the badge. It no longer does the XHR request to get the
number from the server itself.

sidebar-injector.js:
No longer checks the blocklist itself to decide whether to inject into
a tab, assumes that if its injectIntoTab() is called it should go ahead and
inject.

hypothesis-chrome-extension.js:
Calls getBlocklist() each time the user loads a new page, calls updateBadge()
passing in the total from getBlocklist(), doesn't call injectIntoTab() if the
URL is blocked, sets an error instead.

buildext.py no longer copies blocklist.js or the blocklist config setting into
the extension, as they're no longer used.