Change /badge endpoint to /blocklist
And also simplify it: always return the number of annotations, no
longer need to return 0 if the URI is blocked because we're also
returning blocked as a boolean.

This means that it's now the client's responsibility to decide not to
show a badge number on a blocked URI, even though the `total` returned
by the /blocklist endpoint may be > 0.