Fix an issue where WebSocket comms could break if a socket was closed during message processing
If a WebSocket connection was terminated during a gevent-blocking function
call between the `socket.terminated` check and the message actually
being sent via `socket.send()` then socket.send() would throw
an exception and kill the greenlet responsible for iterating
over incoming NSQ messages and dispatching them to appropriate
WebSocket connections.

If this happened, it would consequently break WebSocket
connections for all users until the server was restarted.

This commit is a partial resolution for the problem which
defers the check for a socket being terminated until
_just before_ the socket.send() call so that no blocking
calls will happen inbetween.

This still leaves the possibility of other uncaught exceptions
killing the WS loop.