Validate reset codes with custom type
Using the schema mapping validator to read the reset code and look up
the user in the database has one major disadvantage: the mapping
validator is not called until each individual field in the form
validates. This means that if you submit an incorrect code at the same
time as an incorrect password, you won't be informed that the code is
incorrect.

This commit addresses this problem by creating a custom schema type,
`ResetCode`, which bears the responsibility of transforming a reset code
into a user and back again for the form.