Allow local override of test database
I run PostgreSQL on a different host locally, and have historically
simply changed `conf/test.ini` and then told git to ignore my changes:

    git update-index --assume-unchanged conf/test.ini

This is a bit clunky, and causes problems when moving between branches
that actually have updated the `test.ini` file. So... make it possible
to set a local environment variable that overrides the test database
location.