Fix a broken HypothesisChromeExtension test
The commented-out test was broken. Fix it, and uncomment it.

The fix is to do:

    myFunc['@noCallThru'] = true;

on a function before passing it to proxyquire() as a stub. Due to a bug
in PhantomJS (properties from up the prototype chain are returned when
you do Object.getOwnPropertyNames() on a function in PhantomJS, they
shouldn't be) proxyquire() crashes with a TypeError if you pass it a
function as a stub.

Disabling proxyquire()'s call-through feature for the stubbed function
works around the issue.