Fix a broken blocklist test
The code under test was calling importing the original settings.js
module. proxyquire() does this by default because it merges the keys of
the original module's exported object with those of the proxy object, so
it has to import the original module to do this merging (even if the
proxied object overrides all of those keys). Since settings.js calls
getSettings() at import time, the original getSettings() was getting
called and receiving invalid JSON and crashing.

Setting '@noCallThru' to true disables proxyquire()'s
call-through/merging feature, avoiding the issue.