Make sure the blocklist is always fresh
I discovered that for some reason Chrome does not call onTabUpdated() for the
first URL that's loaded in a new tab (but does call it for the new tab page
when the tab is created, and for subsequent URLs after the first one).

This meant that the HTTP request to the server to ask whether the URL is
blocklisted would not be sent for the first URL loaded in each new tab.
But the _blocklist variable in hypothesis-chrome-extension.js would be defined
and would be holding the Promise from the previous URL (chrome://newtab, from
when the tab was created). So the extension would use this outdated Promise and
would always launch the sidebar on the first page loaded in a tab, even if that
page was blocked.

Even if we could figure out why Chrome does not call onTabUpdated() for the
first URL loaded in each tab and fix that, this makes me realise that there's a
potential race condition when a new URL is loaded between redefining the
_blocklist Promise and trying to activate the sidebar.

So refactor blocklist.js to make sure that it always returns a Promise for the
current URL when asked for one, but only creates a new Promise when necessary
(when the URL has changed).

This also simplifies usage of blocklist.js from
hypothesis-chrome-extension.js's point of view.

Note that this means that if the user refreshes the same page repeatedly it
won't re-send the blocklist request. If the site gets blocked, the user won't
be blocked from launching h on it until they visit another page then come back
to it.