Queue getters need a settings object only
The `get_reader` and `get_writer` functions in `h.queue` only need
access to the application settings, and not the request object.

By changing the signature of these functions to accept a settings object
rather than the entire request, it becomes possible to use these
functions outside of a request context (such as during application
boot). This makes it easier to spin up queue readers outside of the
normal request-response cycle.

The behaviour of the `request.get_queue_reader` and
`request.get_queue_writer` helpers is unchanged by this commit.