Adjust dropdown menu arrow location for groups list
Position the arrow at the top of the groups list
dropdown menu so that it appears beneath the
dropdown indicator in the top-bar.

This is done by hiding the dropdown menu's "built in"
arrow created via CSS :before and :after pseudo-elements
and introducing a separate element which uses the same
technique with borders to create the dropdown arrow.

The use of a separate element for the arrow allows it
to be aligned with the right edge of the dropdown menu's
parent element (in this case, the group name in the top bar)
rather than the dropdown menu's content.