Fix search bar pushing group name when expanded
Override the browser default styling on the search
input's width property so that its maximum
width will be limited by the available space
in the top bar when expanded. Previously it
was inflexible due to the browser default styling
with the group name being pushed off the left side
of the top bar as a result.

There is still an issue at the moment in Firefox
where the share icon gets covered by the search bar
when expanded.

T-90