Fix markdown editor under Angular 1.4
The read-only state was not being passed correctly from
the annotation viewer to the markdown editor, possibly
due to the use of the 'ng-readonly' which in turn
set the 'readonly' attribute that the 'markdown'
directive tried to read.

This converts the markdown editor to use the
preferred style of an element directive plus
a 'read-only' attr.