request.authenticated_user: make sure domain is correct
It's possible  for request.authenticated_userid to contain a userid with
a domain part that doesn't match the h.auth_domain or request.domain
(for example with pre-existing sessions, after a server's h.auth_domain
setting was changed). When this happens request.authenticated_user
should be None.

Also add some tests.