Improve display of dropdown menu for non-groups users
Align the dropdown menu's top arrow with the
bottom arrow of the menu.

As the styling of the dropdown was adjusted for
its new home in the top bar, it became wider and the
previous alignment of the menu resulted in it disappearing
off the left edge of the sidebar.

T-91