Fix the extension killing embedded H
On pages with H embedded, the extension was removing the embed even
when the user did not interact with the extension at all.

Previously, TabState.isTabInactive() would return false for
a tab which H had never been activated on, so in
HypothesisChromeExtension.updateTabDocument() the logic to remove
a tab when de-activating the extension was not triggered.

Recent refactoring of TabState made it, correctly, return
true in TabState.isTabInactive() if H had not been activated,
resulting in updateTabDocument() automatically removing any instance
of H on the page when the tab completed loading.

This commit resolves the issue by introducing an additional
'extensionSidebarInstalled' state flag which tracks whether the
H sidebar has been installed into a tab by the extension.

When a tab state changes, we only remove the sidebar if
extensionSidebarInstalled is true and the tab is now inactive.

This state flag is not persisted, so when reloading the H
state after an extension reload or upgrade, we know that
H needs to be re-installed into a tab if H was previously
active on the tab.

This change also removes the need for the 'force' option
in TabState.setState()