Don't expand URIs if the search URI is a rel="canonical"
If the search URI is known to be a canonical URI for a document, then we
can shortcut URI expansion, and minimise the risk of false positives, by
only searching for the canonical URI. This works because `target.source`
is guaranteed to be set to the canonical URI if one is known when an
annotation is created.

The motivation for this change is the following scenario:

- We have a number of pages -- such as a series of HTML pages for the
  chapters of a published book -- which are marked up with identifying
  metadata.
- Each page contains a [DOI][1] and a <link rel="canonical" ...> tag.
- The DOI is the same on every page.
- The canonical URI is *different* on every page.

This is perfectly plausible and valid, as DOIs are typically assigned at
the publication or article level, not at a page or chapter level.

Prior to this commit, searching on any of these pages (once they had
been annotated) would have had the same result: all the annotations from
all the pages would have been returned from the search query. Why?
Because the DOI is shared between all pages and when we search for URI
"X", we do so in several steps:

1. Expand X into all equivalent URIs known: [A, B, C, ..., X]
2. Normalise each URI: [A', B', C', ..., X']
3. Search for any annotation targeting at least one of these normalised
   URIs.

And, because for each and every page that is annotated an equivalence
is created between the DOI and the canonical URI, the result is a single
equivalence class containing all known URIs (including the DOI).

So, what are our options? We could just not use DOIs, but that's a
change that might have some side effects that are hard to predict.

This commit addresses the problem by noticing that when an annotation is
created on a page which is marked up with a <link rel="canonical"> tag,
its `target.source` property is set to that canonical URI rather than
the page location. This means that we can know for sure that all
annotations for that page can be found by searching for the canonical
URI alone.

Luckily we also store information about whether or not a URI in the
equivalence class was tagged as canonical, which allows us to check if
an incoming search URI is known to be canonical. If it is, we skip step
1 from above and search only by the incoming URI.

Note that this approach does not change what happens if you search for
the DOI, or even for a non-canonical URI known for one of the pages. In
this case, all annotations across all pages will still be returned.

[1]: https://en.wikipedia.org/wiki/Digital_object_identifier

Fixes #2634.
Fixes #2651.