Fix extension not being installed if a tab is loaded via pre-rendering
In certain scenarios, Chrome will pre-load tabs in the background
when it predicts that the user will navigate to a particular URL.

When the user does then perform that navigation, it replaces the
current tab with the pre-loaded tab. The pre-loaded tab has
a different tab ID and the extension is notified via an
`onReplace` event rather than `onUpdated`.

 * Handle this case in the Chrome extension by updating the
   internal H state of the tab in TabState accordingly.

 * This commit also fixes a potential attempted double-installation
   of H on a tab by setting the flag indicating whether the sidebar
   has been installed optimistically _just before_ starting the injector
   rather than after injection completes.

 * Update the browser action badge to reflect the count of
   annotations on a URL when a navigation begins rather than waiting
   for navigation to fully complete, something that only happens
   once _all_ resources on a page have loaded.

   A consequence is that the H badge now feels much more responsive.