Set up a constraint naming convention for sqlalchemy
In order to be able to add and remove constraints using alembic, they
must be named. Luckily for us PostgreSQL does actually have a reasonably
sensible constraint naming convention, but enforcing this in sqlalchemy
allows us to know what constraints are called without referring to the
database in question.

In addition, the default PostgreSQL naming convention suffers from
clashes: an FK constraint on the column "id" on a join table called
"foo_bar" will be named "foo_bar_id_fkey", which is the same name as
would be given to an FK constraint on the "bar_id" column on table
"foo". The naming convention chosen here doesn't suffer from the same
problem (presuming that you never use "__" in column names).

This commit includes a migration to update existing constraints to the
new naming convention.