Turn on pyramid's conditional_response machinery
For many responses, it can be useful to generate an ETag which is a hash
of their contents. This can be stored by user-agents, which can then
send an If-None-Match header with a subsequent request for the same
resource. In combination with appropriate Cache-Control headers, this
should allow us to save some bandwidth and speed up the client.