Add a basic categorized logging facility
This adds a minimal categorized logging facility and
adds a facility to use it for logging web socket
messages that are received.

This serves two purposes:

 * Tagging logging messages with the component they
   relate to.

 * Setting the logging level for different parts of
   the app at runtime via `window.H_DEBUG.logging`

   eg.
     window.H_DEBUG.logging.setLevel('debug')

   Will turn on debug logging for all loggers.

 * In future we may extend this to report metrics and
   log messages etc. to the server.

At the moment, this is a minimal custom implementation.
In future if we decide we do need more features, it should
be possible to swap out an existing well known service such as
Winston or Bunyan with very few changes to the users of the logging
service.