Deliver up-to-date groups list in groups leave/join notifications
When pushing a session state change notification to clients
after joining or leaving a group, the session model could
be out of date because the groups view published the notification
prior to committing the database transaction.

If the websocket server picked up the message from NSQ and queried
the DB for the session state prior to the commit happening,
the client would receive a notification but the payload
would be out of date.

We considered two possible approaches to fix this - one
was to commit the transaction in the groups views before
publishing the notification, the other was to serialize the
session model as part of the notification.

This commit takes the latter approach for consistency with how
annotation notifications work and also as slightly more
explicit to read. We may want to reconsider in future.

Fixes #2671