Fix groups list not updating after initial login
When the <groups-list> directive is instantiated,
login is not complete and the groups list contains only
the Public group.

Once the login completes, the groups list was updated
because GROUPS_CHANGED was not broadcast for the initial
groups load. The reason for this was to avoid an unnecessary
route reload in `AppController`.

Resolve this by always firing the `SESSION_CHANGED`,
`USER_CHANGED` and `GROUPS_CHANGED` events when
they change, but include an additional piece of data
indicating whether this is the first load or not. This is used
in `AppController` to avoid an unnecessary route reload.